cmd: fold the schema-snapshot unit tests into one file
This branch added two unit test files to cmd for the same feature:
snapshotStartReason_test.go (the start- and exit-reason classifiers) and
schemaSnapshotIntervalFlag_test.go (interval-flag validation). Merge them
into schemaSnapshotCapture_test.go, matching the one-file-per-feature
pattern the other unit tests in cmd already follow (analyzeSchema_test.go,
conflictDetectionCache_test.go, exportDataStatus_test.go).

No pre-existing unit test file was a sensible host: there is none for
exportData.go or export.go, and cmd/integration_test.go carries real
Oracle/resumption tests rather than being a shared harness.

Test bodies are unchanged. One stale doc comment is fixed on the way:
TestExportDataExitReason claimed the reason was shared with "the inline
`return false` error paths", which have not existed since the single
exit-capture defer landed.
