+++ yb-voyager/cmd/schemaSnapshotCapture_test.go
@@ -20,7 +20,9 @@ package cmd
 import (
 	"testing"
 
+	"github.com/spf13/cobra"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"github.com/yugabyte/yb-voyager/yb-voyager/src/schemasnapshot"
 )
@@ -49,11 +51,10 @@ func TestSnapshotStartReasonFor(t *testing.T) {
 	}
 }
 
-// TestExportDataExitReason pins the abnormal-exit reason classification. This is
-// shared by the atexit hook AND the inline `return false` error paths: a signal
-// kills the in-flight dump/streaming child, so the export reports failure and hits
-// the inline path first. Verified end-to-end against PostgreSQL — before this was
-// shared, SIGINT/SIGTERM/SIGUSR2 all recorded reason=error.
+// TestExportDataExitReason pins the abnormal-exit reason classification, which is shared
+// by the atexit hook and by exportData's exit defer. Verified end-to-end against
+// PostgreSQL — before the two shared this, SIGINT/SIGTERM/SIGUSR2 all recorded
+// reason=error.
 func TestExportDataExitReason(t *testing.T) {
 	origShutdown, origEndMigration := ProcessShutdownRequested.Load(), EndMigrationStopRequested.Load()
 	t.Cleanup(func() {
@@ -80,6 +81,59 @@ func TestExportDataExitReason(t *testing.T) {
 			ProcessShutdownRequested.Store(tt.shutdownReq)
 			EndMigrationStopRequested.Store(tt.endMigrationStop)
 			assert.Equal(t, tt.want, exportDataExitReason())
+		})
+	}
+}
+
+// TestValidateSchemaSnapshotCaptureInterval pins the flag validation. An interval of 0
+// or less would silently disable periodic capture, so it must fail at startup instead.
+//
+// The last case pins the flag-presence scoping: with the flag absent from the command,
+// the guard no-ops whatever the global holds. It sets 0 explicitly to exercise that,
+// which a real `export data from target` run would not -- IntVar writes 60 into the
+// global at registration, so that scoping is defensive rather than load-bearing.
+func TestValidateSchemaSnapshotCaptureInterval(t *testing.T) {
+	orig := schemaSnapshotCaptureInterval
+	t.Cleanup(func() { schemaSnapshotCaptureInterval = orig })
+
+	withFlag := func() *cobra.Command {
+		cmd := &cobra.Command{Use: "export-data-stub"}
+		registerSchemaSnapshotIntervalFlag(cmd)
+		return cmd
+	}
+
+	tests := []struct {
+		name     string
+		cmd      func() *cobra.Command
+		interval int
+		wantErr  bool
+	}{
+		{"default interval passes", withFlag, 60, false},
+		{"the minimum interval passes", withFlag, 1, false},
+		{"zero is rejected", withFlag, 0, true},
+		{"negative is rejected", withFlag, -5, true},
+		{
+			"a command without the flag is not validated (export data from target)",
+			func() *cobra.Command { return &cobra.Command{Use: "export-data-from-target-stub"} },
+			0,
+			false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			cmd := tt.cmd()
+			// registerSchemaSnapshotIntervalFlag binds the global and resets it to the
+			// flag default, so set the value under test after building the command.
+			schemaSnapshotCaptureInterval = tt.interval
+
+			err := validateSchemaSnapshotCaptureInterval(cmd)
+			if tt.wantErr {
+				require.Error(t, err, "an interval of %d must be rejected", tt.interval)
+				assert.Contains(t, err.Error(), "must be at least 1",
+					"the error must say what a valid interval is")
+				return
+			}
+			require.NoError(t, err)
 		})
 	}
 }
